+++ lib/rdf/datatypes/decimal.ex
@@ -110,4 +110,7 @@ defmodule RDF.Decimal do
   @impl RDF.Datatype
   def equal_value?(left, right), do: RDF.Numeric.equal_value?(left, right)
 
+  @impl RDF.Datatype
+  def compare(left, right), do: RDF.Numeric.compare(left, right)
+
 end