+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - `RDF.Description.values/1`, `RDF.Graph.values/1`, `RDF.Dataset.values/1` and
 	`RDF.Data.values/1` returning a map of `RDF.Term.value/1` converted native 
 	Elixir values from the respective structure of RDF terms
+- `RDF.Literal.compare/2`, `RDF.Literal.less_than?/2` and `RDF.Literal.greater_than?/2`  
+  for `RDF.Datatype` aware comparisons of `RDF.Literal`s  
 
 
 ### Fixed